refactor: hide JSON configuration factory
