+++ src/config/factory.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace blackhole {
+inline namespace v1 {
+namespace config {
+
+template<typename T>
+class factory;
+
+}  // namespace config
+}  // namespace v1
+}  // namespace blackhole
